skills_scraper/requirements.txt
python-dotenv==0.19.1
Scrapy==1.7.3
